always wait for first message regardless of title prefix setting
Discord will literally block the bot for responding to a thread before the message is even sent lol, since apparently that's a thing. Probably an exploit hole on Discord's end (what if the user sends the thread-create but never the message? Is the thread locked in unacccesible state?) but that's their problem not mine.